# Further perf improvement to GetTagHelpersDelta