I noticed this method in a completion performance trace I was looking at and saw Dustin made an O(n^2) => O(n) improvement in this code that likely addressed the performance issue I was seeing (PR: https://github.com/dotnet/razor/pull/9086). However, while admiring his awesome fix, I did notice that it could be slightly optimized and get close to a 50% reduction in most cases by observing the fact that we most likely don't need to call TagHelperData.Compute twice.

Benchmark results look promising:

```
|                                        Method |        Job |            Toolchain | ReduceComputeCalls |         Mean |       Error |      StdDev |       Median |          Min |         Max | Allocated |
|---------------------------------------------- |----------- |--------------------- |------------------- |-------------:|------------:|------------:|-------------:|-------------:|------------:|----------:|
|               'Calculate Delta - New project' | Job-ZGGOIK | .NET Framework 4.7.2 |              False |    10.326 us |   0.5618 us |   1.5660 us |    10.450 us |     7.100 us |    14.70 us |    8192 B |
|            'Calculate Delta - Remove project' | Job-ZGGOIK | .NET Framework 4.7.2 |              False |     8.694 us |   0.3199 us |   0.9230 us |     8.800 us |     6.500 us |    11.10 us |    8192 B |
|    'Calculate Delta - Add lots of TagHelpers' | Job-ZGGOIK | .NET Framework 4.7.2 |              False | 2,235.087 us | 154.2317 us | 454.7556 us | 2,448.300 us | 1,430.000 us | 3,028.50 us |  238008 B |
| 'Calculate Delta - Remove lots of TagHelpers' | Job-ZGGOIK | .NET Framework 4.7.2 |              False | 1,388.991 us |  88.3521 us | 260.5082 us | 1,477.450 us |   874.250 us | 1,907.75 us |  117792 B |
|     'Calculate Delta - Mutate two TagHelpers' | Job-ZGGOIK | .NET Framework 4.7.2 |              False | 1,959.338 us | 107.5459 us | 315.4134 us | 2,065.800 us | 1,240.800 us | 2,376.90 us |  148736 B |
|                 'Calculate Delta - No change' | Job-ZGGOIK | .NET Framework 4.7.2 |              False | 1,909.361 us | 114.3374 us | 337.1265 us | 2,058.250 us | 1,273.500 us | 2,430.30 us |  148736 B |
|               'Calculate Delta - New project' | Job-ZGGOIK | .NET Framework 4.7.2 |               True |    10.273 us |   0.4886 us |   1.3862 us |    10.100 us |     7.400 us |    14.10 us |    8192 B |
|            'Calculate Delta - Remove project' | Job-ZGGOIK | .NET Framework 4.7.2 |               True |     8.673 us |   0.4254 us |   1.2409 us |     8.700 us |     5.950 us |    11.85 us |    8192 B |
|    'Calculate Delta - Add lots of TagHelpers' | Job-ZGGOIK | .NET Framework 4.7.2 |               True | 1,313.668 us |  83.4186 us | 240.6818 us | 1,370.700 us |   771.500 us | 1,713.60 us |   86656 B |
| 'Calculate Delta - Remove lots of TagHelpers' | Job-ZGGOIK | .NET Framework 4.7.2 |               True |   817.705 us |  45.5482 us | 132.8663 us |   831.300 us |   503.400 us | 1,053.10 us |   47520 B |
|     'Calculate Delta - Mutate two TagHelpers' | Job-ZGGOIK | .NET Framework 4.7.2 |               True | 1,850.599 us | 120.5240 us | 355.3676 us | 2,016.250 us | 1,211.200 us | 2,305.40 us |  148736 B |
|                 'Calculate Delta - No change' | Job-ZGGOIK | .NET Framework 4.7.2 |               True | 1,085.055 us |  81.7935 us | 241.1699 us | 1,134.050 us |   666.100 us | 1,633.20 us |   78464 B |
```